// src/shared/lib/models/Form/Validation/ValidationCode.ts
export enum ValidationCode {
  REQUIRED = 'required',
  EMAIL = 'email',
  EMAIL_RFC5322 = 'email_rfc5322',
  PHONE_INTERNATIONAL = 'phone_international',
  NUMBER = 'number',
  MIN = 'min',
  MAX = 'max',
  BETWEEN = 'between',
  COLOR_HEX = 'color_hex',
  HTTP_URL = 'http_url',
  HTTP_HEADER_KEY = 'http_header_key',
  PRINTABLE_ASCII = 'printable_ascii',
}